+++ modules/actor/sheet/cpr-container-sheet.js
@@ -41,21 +41,25 @@ export default class CPRContainerActorSheet extends CPRActorSheet {
    */
   getData() {
     LOGGER.trace("getData | CPRContainerSheet | Called.");
-    const data = super.getData();
-    data.isGM = game.user.isGM;
-    data.userOwnedActors = game.actors.filter((a) => a.isOwner && a.type === "character");
-    data.userOwnedActors.unshift({ id: "", name: "--" });
-    if (game.user.character !== undefined) {
-      data.userCharacter = game.user.character.id;
-      if (this.tradePartnerId === undefined) {
-        this.tradePartnerId = game.user.character.id;
+    const foundryData = super.getData();
+    const cprActorData = foundryData.actor.system;
+
+    cprActorData.userOwnedActors = game.actors.filter((a) => a.isOwner && a.type === "character");
+    cprActorData.userOwnedActors.unshift({ id: "", name: "--" });
+    if (game.user.character !== undefined && game.user.character !== null) {
+      cprActorData.userCharacter = game.user.character.id;
+      if (!cprActorData.tradePartnerId) {
+        cprActorData.tradePartnerId = game.user.character.id;
       }
-      data.tradePartnerId = this.tradePartnerId;
+      if (this.tradePartnerId) {
+        cprActorData.tradePartnerId = this.tradePartnerId;
+      }
     } else {
-      data.userCharacter = "";
-      data.tradePartnerId = this.tradePartnerId;
+      cprActorData.userCharacter = "";
+      cprActorData.tradePartnerId = this.tradePartnerId;
     }
-    return data;
+    foundryData.data.system = cprActorData;
+    return foundryData;
   }
 
   /**
@@ -160,7 +164,7 @@ export default class CPRContainerActorSheet extends CPRActorSheet {
       // Only update if we aren't deleting the item.  Item deletion is handled in this._deleteOwnedItem()
       // The same holds for purchasing the item, as it is handled by this._purchaseItem()
       if (actionType !== "delete" && actionType !== "purchase") {
-        this.actor.updateEmbeddedDocuments("Item", [{ _id: item.id, data: item.data.data }]);
+        this.actor.updateEmbeddedDocuments("Item", [{ _id: item.id, data: item.system }]);
       }
     }
   }
@@ -177,11 +181,11 @@ export default class CPRContainerActorSheet extends CPRActorSheet {
   _renderItemCard(event) {
     LOGGER.trace("_renderItemCard | CPRContainerSheet | Called.");
     const itemId = CPRActorSheet._getItemId(event);
-    const item = this.actor.items.find((i) => i.data._id === itemId);
+    const item = this.actor.items.find((i) => i._id === itemId);
     if (event.ctrlKey) {
       CPRChat.RenderItemCard(item);
     } else {
-      const playersCanModify = getProperty(this.actor.data, "flags.cyberpunk-red-core.players-modify");
+      const playersCanModify = getProperty(this.actor, "flags.cyberpunk-red-core.players-modify");
       if (playersCanModify || game.user.isGM) {
         item.sheet.render(true, { editable: true });
       } else {
@@ -224,13 +228,13 @@ export default class CPRContainerActorSheet extends CPRActorSheet {
       return;
     }
 
-    const transferredItemData = duplicate(item.data);
+    const transferredItemData = duplicate(item);
     let cost = 0;
-    if (item.type === "ammo" && item.data.data.variety !== "grenade" && item.data.data.variety !== "rocket") {
+    if (item.type === "ammo" && item.system.variety !== "grenade" && item.system.variety !== "rocket") {
       // Ammunition, which is neither grenades nor rockets, are prices are for 10 of them (pg. 344)
-      cost = item.data.data.price.market / 10;
+      cost = item.system.price.market / 10;
     } else {
-      cost = item.data.data.price.market;
+      cost = item.system.price.market;
     }
     if (!all) {
       const itemText = SystemUtils.Format("CPR.dialog.purchasePart.text", { itemName: item.name });
@@ -244,18 +248,18 @@ export default class CPRContainerActorSheet extends CPRActorSheet {
         SystemUtils.DisplayMessage("warn", SystemUtils.Localize("CPR.dialog.purchasePart.wrongAmountWarning"));
         return;
       }
-      transferredItemData.data.amount = newAmount;
+      transferredItemData.system.amount = newAmount;
       cost *= newAmount;
     } else {
       cost *= (typeof item.data.data.amount !== "undefined") ? parseInt(item.data.data.amount, 10) : 1;
     }
     const tradePartnerActor = game.actors.get(this.tradePartnerId);
-    if (!getProperty(this.actor.data, "flags.cyberpunk-red-core.items-free")) {
-      if (tradePartnerActor.data.data.wealth.value < cost) {
+    if (!getProperty(this.actor, "flags.cyberpunk-red-core.items-free")) {
+      if (tradePartnerActor.system.wealth.value < cost) {
         SystemUtils.DisplayMessage("warn", SystemUtils.Localize("CPR.messages.tradePriceWarn"));
         return;
       }
-      const { amount } = transferredItemData.data;
+      const { amount } = transferredItemData.system;
       const username = game.user.name;
       let reason = "";
       if (amount > 1) {
@@ -284,12 +288,12 @@ export default class CPRContainerActorSheet extends CPRActorSheet {
     if (tradePartnerActor.automaticallyStackItems(new CPRItem(transferredItemData))) {
       await tradePartnerActor.createEmbeddedDocuments("Item", [transferredItemData]);
     }
-    if (!getProperty(this.actor.data, "flags.cyberpunk-red-core.infinite-stock")) {
+    if (!getProperty(this.actor, "flags.cyberpunk-red-core.infinite-stock")) {
       if (all) {
         await this._deleteOwnedItem(item, true);
       } else {
-        const keepAmount = item.data.data.amount - transferredItemData.data.amount;
-        await this.actor.updateEmbeddedDocuments("Item", [{ _id: item.id, "data.amount": keepAmount }]);
+        const keepAmount = item.system.amount - transferredItemData.system.amount;
+        await this.actor.updateEmbeddedDocuments("Item", [{ _id: item.id, "system.amount": keepAmount }]);
       }
     }
   }
@@ -313,51 +317,51 @@ export default class CPRContainerActorSheet extends CPRActorSheet {
     const dragData = JSON.parse(event.dataTransfer.getData("text/plain"));
     const tradePartnerActor = game.actors.get(dragData.actorId);
 
-    const itemData = duplicate(dragData.data);
-    const amount = itemData.data.amount ? parseInt(itemData.data.amount, 10) : 1;
-    const vendorData = this.actor.data;
-    const vendorConfig = vendorData.data.vendor;
+    const item = duplicate(dragData.data);
+    const cprItemData = item.system;
+    let cprItemName = dragData.name;
+    const amount = cprItemData.amount ? parseInt(cprItemData.amount, 10) : 1;
+    const vendorData = this.actor.system;
+    const vendorConfig = vendorData.vendor;
     const username = game.user.name;
 
     let cost = 0;
 
-    if (itemData.type === "weapon") {
-      const { ammoId } = itemData.data.magazine;
+    if (item.type === "weapon") {
+      const { ammoId } = cprItemData.magazine;
       if (ammoId !== "") {
         SystemUtils.DisplayMessage("warn", SystemUtils.Localize("CPR.messages.tradeLoadedWeaponWarn"));
         return;
       }
     }
 
-    if (itemData.type === "ammo" && itemData.data.variety !== "grenade" && itemData.data.variety !== "rocket") {
+    if (item.type === "ammo" && cprItemData.variety !== "grenade" && cprItemData.variety !== "rocket") {
       // Ammunition, which is neither grenades nor rockets, are prices are for 10 of them (pg. 344)
-      cost = parseInt(parseInt(itemData.data.price.market, 10) / 10, 10);
+      cost = parseInt(parseInt(cprItemData.price.market, 10) / 10, 10);
     } else {
-      cost = parseInt(itemData.data.price.market, 10);
+      cost = parseInt(cprItemData.price.market, 10);
     }
-    const percent = parseInt(vendorConfig.itemTypes[itemData.type].purchasePercentage, 10);
+    const percent = parseInt(vendorConfig.itemTypes[item.type].purchasePercentage, 10);
 
-    let itemName = itemData.name;
-
-    if (itemData.data.isUpgraded) {
-      itemData.data.upgrades.forEach((upgrade) => {
-        const upgradeItem = tradePartnerActor.items.find((i) => i.data._id === upgrade._id);
+    if (cprItemData.isUpgraded) {
+      cprItemData.upgrades.forEach((upgrade) => {
+        const upgradeItem = tradePartnerActor.items.find((i) => i._id === upgrade._id);
         if (upgradeItem) {
-          cost += upgradeItem.data.data.price.market;
+          cost += upgradeItem.system.price.market;
         }
       });
-      itemName = `${SystemUtils.Localize("CPR.global.generic.upgraded")} ${itemName}`;
+      cprItemName = `${SystemUtils.Localize("CPR.global.generic.upgraded")} ${cprItemName}`;
     }
 
     let vendorOffer = parseInt(((amount * cost * percent) / 100), 10);
-    vendorOffer = Math.min(vendorOffer, vendorData.data.wealth.value);
+    vendorOffer = Math.min(vendorOffer, vendorData.wealth.value);
 
     const offerMessage = `${SystemUtils.Format(
       "CPR.dialog.container.vendor.offerToBuy",
       {
         vendorName: tradePartnerActor.name,
         vendorOffer,
-        itemName,
+        cprItemName,
         percent,
       },
     )}`;
@@ -366,31 +370,35 @@ export default class CPRContainerActorSheet extends CPRActorSheet {
     if (formData !== undefined) {
       let createItems = [];
       const deleteItems = [];
-      if (itemData.data.isUpgraded) {
-        itemData.data.upgrades.forEach(async (upgrade) => {
-          const upgradeItem = tradePartnerActor.items.find((i) => i.data._id === upgrade._id);
+      if (cprItemData.isUpgraded) {
+        cprItemData.upgrades.forEach(async (upgrade) => {
+          const upgradeItem = tradePartnerActor.items.find((i) => i._id === upgrade._id);
           if (upgradeItem) {
-            const newItem = duplicate(upgradeItem.data);
-            newItem.data.isInstalled = false;
-            newItem.data.install = "";
+            const newItem = duplicate(upgradeItem);
+            newItem.system.isInstalled = false;
+            newItem.system.install = "";
             createItems.push(newItem);
-            deleteItems.push(upgradeItem.data._id);
+            deleteItems.push(upgradeItem._id);
           }
         });
-        itemData.data.isUpgraded = false;
-        itemData.data.upgrades = [];
+        cprItemData.isUpgraded = false;
+        cprItemData.upgrades = [];
       }
 
-      createItems.push(itemData);
-      deleteItems.push(itemData._id);
+      createItems.push({
+        name: item.name,
+        system: cprItemData,
+        type: item.type,
+      });
+      deleteItems.push(item._id);
 
-      const infiniteStock = getProperty(this.actor.data, "flags.cyberpunk-red-core.infinite-stock");
+      const infiniteStock = getProperty(this.actor, "flags.cyberpunk-red-core.infinite-stock");
 
       if (infiniteStock) {
         const itemList = createItems;
-        itemList.forEach((item) => {
-          if (this.actor.items.find((i) => i.data.name === item.name)) {
-            createItems = createItems.filter((ci) => ci._id !== item._id);
+        itemList.forEach((infiniteItem) => {
+          if (this.actor.items.find((i) => i.name === infiniteItem.name)) {
+            createItems = createItems.filter((ci) => ci._id !== infiniteItem._id);
           }
         });
       }
@@ -406,7 +414,7 @@ export default class CPRContainerActorSheet extends CPRActorSheet {
           "CPR.containerSheet.tradeLog.multipleSold",
           {
             amount,
-            name: itemData.name,
+            name: item.name,
             price: vendorOffer,
             vendor: this.actor.name,
           },
@@ -414,14 +422,14 @@ export default class CPRContainerActorSheet extends CPRActorSheet {
       } else {
         reason = `${SystemUtils.Format(
           "CPR.containerSheet.tradeLog.singleSold",
-          { name: itemData.name, price: vendorOffer, vendor: this.actor.name },
+          { name: item.name, price: vendorOffer, vendor: this.actor.name },
         )} - ${username}`;
       }
       const vendorReason = `${SystemUtils.Format(
         "CPR.containerSheet.tradeLog.vendorPurchased",
         {
-          name: itemData.name,
-          quantity: itemData.data.amount,
+          name: item.name,
+          quantity: cprItemData.amount,
           seller: tradePartnerActor.name,
           price: vendorOffer,
         },
@@ -497,12 +505,12 @@ export default class CPRContainerActorSheet extends CPRActorSheet {
    */
   async _onDrop(event) {
     LOGGER.trace("_onDrop | CPRContainerSheet | Called.");
-    const playersCanCreate = getProperty(this.actor.data, "flags.cyberpunk-red-core.players-create");
-    const playersCanSell = getProperty(this.actor.data, "flags.cyberpunk-red-core.players-sell");
+    const playersCanCreate = getProperty(this.actor, "flags.cyberpunk-red-core.players-create");
+    const playersCanSell = getProperty(this.actor, "flags.cyberpunk-red-core.players-sell");
     if (game.user.isGM || playersCanCreate || playersCanSell) {
       if (!game.user.isGM && !playersCanCreate) {
         const dragData = JSON.parse(event.dataTransfer.getData("text/plain"));
-        const vendorData = this.actor.data.data.vendor;
+        const vendorData = this.actor.system.vendor;
         if (dragData.type === "Item") {
           const itemData = dragData.data;
           if (typeof vendorData.itemTypes[itemData.type] !== "undefined" && vendorData.itemTypes[itemData.type].isPurchasing) {
@@ -527,7 +535,7 @@ export default class CPRContainerActorSheet extends CPRActorSheet {
    */
   async _configureSellTo() {
     LOGGER.trace("_configureSellTo | CPRContainerSheet | Called.");
-    const actorData = duplicate(this.actor.data);
+    const cprActorData = duplicate(this.actor.system);
     const promptData = {};
     promptData.itemTypes = [];
     const itemEntities = game.system.template.Item;
@@ -536,7 +544,7 @@ export default class CPRContainerActorSheet extends CPRActorSheet {
         promptData.itemTypes.push(itemType);
       }
     });
-    promptData.currentConfig = actorData.data.vendor;
+    promptData.currentConfig = cprActorData.vendor;
 
     promptData.itemTypes.forEach((itemType) => {
       if (typeof promptData.currentConfig.itemTypes[itemType] === "undefined") {
@@ -554,8 +562,8 @@ export default class CPRContainerActorSheet extends CPRActorSheet {
         const purchasePercentage = (isPurchasing) ? formData[`pct-${itemType}`] : 0;
         promptData.currentConfig.itemTypes[itemType] = { isPurchasing, purchasePercentage };
       });
-      setProperty(actorData, "data.vendor", promptData.currentConfig);
-      this.actor.update(actorData);
+      setProperty(cprActorData, "data.vendor", promptData.currentConfig);
+      this.actor.update(cprActorData);
     }
   }