+++ modules/dialog/cpr-split-item-prompt.js
@@ -1,4 +1,4 @@
-/* global renderTemplate, FormDataExtended, Dialog */
+/* global renderTemplate, FormDataExtended, Dialog foundry */
 import SystemUtils from "../utils/cpr-systemUtils.js";
 import LOGGER from "../utils/cpr-logger.js";
 
@@ -14,7 +14,8 @@ export default class SplitItemPrompt {
         // eslint-disable-next-line no-shadow
         const _onConfirm = (html) => {
           LOGGER.trace("_onConfirm | Dialog SplitItemPrompt | called.");
-          const formData = new FormDataExtended(html.find("form")[0]).toObject();
+          const fd = new FormDataExtended(html.find("form")[0]);
+          const formData = foundry.utils.expandObject(fd.object);
           resolve(formData);
         };
         new Dialog({