+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Version DEV | Date: WIP
 
 **New Features**
--
+- Foundry V10 Compatibility
 
 **Changes**
 -